backend/consumer/requirements.txt
confluent-kafka==2.3.0
pandas==2.2.2
python-dotenv==1.0.1